requirements.txt
aiohttp>=3.8.0
aiofiles>=0.8.0
beautifulsoup4>=4.10.0
configparser>=5.0.2
markdown>=3.3.6
numpy>=1.20.0
openai>=1.0.0
Pillow>=9.0.0
python-dateutil>=2.8.2
pytz>=2021.1
requests>=2.26.0
tenacity>=8.0.1
WeasyPrint>=54.0  # Optional for PDF generation
anthropic>=0.5.0  # Optional for Claude API
matplotlib>=3.5.0  # For chart generation
uuid>=1.30  # For UUID generation in OPC collector
fnmatch>=0.1.0  # For pattern matching in data curator
base64io>=1.0.3  # For efficient base64 encoding/decoding
aiosignal>=1.2.0  # Required by aiohttp
async-timeout>=4.0.0  # Required for async operations